+++ client/src/app/modules/crag/crag-list/crag-list.component.ts
@@ -2,7 +2,7 @@ import { Component, DestroyRef, inject, OnInit } from '@angular/core';
 import { Crag } from '../../../models/crag';
 import { CragsService } from '../../../services/crud/crags.service';
 import { LoadingState } from '../../../enums/loading-state';
-import { ConfirmationService, PrimeIcons, SelectItem } from 'primeng/api';
+import { ConfirmationService } from 'primeng/api';
 import { TranslocoDirective, TranslocoService } from '@jsverse/transloco';
 import { marker } from '@jsverse/transloco-keys-manager/marker';
 import { forkJoin, Observable } from 'rxjs';
@@ -10,6 +10,12 @@ import { select, Store } from '@ngrx/store';
 import { selectIsMobile } from '../../../ngrx/selectors/device.selectors';
 import { DialogService, DynamicDialogRef } from 'primeng/dynamicdialog';
 import { OrderItemsComponent } from '../../shared/components/order-items/order-items.component';
+import {
+  buildTopoListOrderDirectionOptions,
+  buildTopoListOrderOptions,
+  topoListSortFieldAndOrder,
+  TopoListSortSelectOption,
+} from '../../../utility/topo-list-sort';
 
 import { DataView } from 'primeng/dataview';
 import { Select } from 'primeng/select';
@@ -61,8 +67,10 @@ export class CragListComponent implements OnInit {
   public crags: Crag[];
   public loading = LoadingState.LOADING;
   public loadingStates = LoadingState;
-  public sortOptions: SelectItem[];
-  public sortKey: SelectItem;
+  public orderOptions: TopoListSortSelectOption[];
+  public orderKey: TopoListSortSelectOption;
+  public orderDirectionOptions: TopoListSortSelectOption[];
+  public orderDirectionKey: TopoListSortSelectOption;
   public sortOrder: number;
   public sortField: string;
   public isMobile$: Observable<boolean>;
@@ -89,45 +97,12 @@ export class CragListComponent implements OnInit {
     forkJoin([this.cragsService.getCrags()]).subscribe(([crags]) => {
       this.crags = crags;
       this.loading = LoadingState.DEFAULT;
-      this.sortOptions = [
-        {
-          icon: PrimeIcons.SORT_AMOUNT_DOWN_ALT,
-          label: this.translocoService.translate(marker('sortAscending')),
-          value: '!orderIndex',
-        },
-        {
-          icon: PrimeIcons.SORT_AMOUNT_DOWN,
-          label: this.translocoService.translate(marker('sortDescending')),
-          value: 'orderIndex',
-        },
-        {
-          icon: PrimeIcons.SORT_ALPHA_DOWN,
-          label: this.translocoService.translate(marker('sortAZ')),
-          value: '!name',
-        },
-        {
-          icon: 'pi pi-sort-alpha-down-alt',
-          label: this.translocoService.translate(marker('sortZA')),
-          value: 'name',
-        },
-      ];
-      this.sortKey = this.sortOptions[0];
+      this.initSortControls();
     });
   }
 
-  /**
-   * Sets the sort field and order.
-   * @param event Sort change event.
-   */
-  onSortChange(event: any) {
-    const value = event.value.value;
-    if (value.indexOf('!') === 0) {
-      this.sortOrder = 1;
-      this.sortField = value.substring(1, value.length);
-    } else {
-      this.sortOrder = -1;
-      this.sortField = value;
-    }
+  onSortControlsChange() {
+    this.applySort();
   }
 
   /**
@@ -150,5 +125,27 @@ export class CragListComponent implements OnInit {
     this.ref.onClose.pipe(takeUntilDestroyed(this.destroyRef)).subscribe(() => {
       this.refreshData();
     });
+  }
+
+  private initSortControls() {
+    this.orderOptions = buildTopoListOrderOptions(this.translocoService);
+    this.orderDirectionOptions = buildTopoListOrderDirectionOptions(
+      this.translocoService,
+    );
+    this.orderKey = this.orderOptions[0];
+    this.orderDirectionKey = this.orderDirectionOptions[1];
+    this.applySort();
+  }
+
+  private applySort() {
+    if (!this.orderKey || !this.orderDirectionKey) {
+      return;
+    }
+    const { sortField, sortOrder } = topoListSortFieldAndOrder(
+      this.orderKey,
+      this.orderDirectionKey,
+    );
+    this.sortField = sortField;
+    this.sortOrder = sortOrder;
   }
 }